!Project Files/Software/Wabash Interface/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "wabash-interface"
version = "0.1.0"
description = "Desktop user interface for Wabash capstone transmitter/receiver system"
requires-python = ">=3.11"
dependencies = [
  "pyserial>=3.5",
]

[project.scripts]
wabash-interface = "wabash_interface.main:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
